moved partialEvaluation, first version of select-based grounding
