+++ benchmarks/stratified_evaluation/test-results/3col-10-18.clingo.out
@@ -1,0 +1,13 @@
+clingo version 5.2.2
+Reading from test-src/3col-10-18.asp
+Solving...
+Answer: 1
+node(0) node(1) node(2) node(3) node(4) node(5) node(6) node(7) node(8) node(9) link(0,2) link(2,0) link(1,5) link(5,1) link(4,5) link(5,4) link(7,8) link(8,7) link(0,4) link(4,0) link(3,8) link(8,3) link(1,3) link(3,1) link(1,8) link(8,1) link(2,6) link(6,2) link(2,9) link(9,2) link(3,6) link(6,3) link(2,3) link(3,2) link(1,7) link(7,1) link(2,7) link(7,2) link(5,8) link(8,5) link(6,7) link(7,6) link(5,9) link(9,5) link(0,3) link(3,0) green(0) green(1) blue(2) red(3) blue(4) red(5) green(6) red(7) blue(8) green(9)
+SATISFIABLE
+
+Models       : 1+
+Calls        : 1
+Time         : 0.001s (Solving: 0.00s 1st Model: 0.00s Unsat: 0.00s)
+CPU Time     : 0.001s
+Command exited with non-zero status 10
+TIME-STATS	clingo -n 1 test-src/3col-10-18.asp	0.00	10
