+++ benchmarks/stratified_evaluation/test-results/3col-20-38.alpha.out
@@ -1,0 +1,6 @@
+  465 I tuwien.kr.alpha.Main Not writing dependency or component graphs, starting preprocessing... 
+  476 I uwien.kr.alpha.Alpha Preprocessing InternalProgram! 
+Answer set 1:
+{ blue(0), blue(4), blue(6), blue(9), blue(14), blue(16), blue(19), green(2), green(3), green(5), green(7), green(13), green(18), link(0, 3), link(0, 8), link(0, 10), link(0, 13), link(0, 15), link(0, 18), link(1, 6), link(1, 18), link(2, 11), link(3, 0), link(3, 9), link(3, 15), link(3, 16), link(3, 17), link(4, 5), link(4, 7), link(4, 13), link(5, 4), link(5, 9), link(5, 10), link(5, 15), link(6, 1), link(6, 8), link(6, 10), link(6, 12), link(6, 13), link(6, 15), link(7, 4), link(7, 8), link(7, 9), link(7, 11), link(8, 0), link(8, 6), link(8, 7), link(8, 18), link(9, 3), link(9, 5), link(9, 7), link(9, 15), link(10, 0), link(10, 5), link(10, 6), link(10, 13), link(10, 14), link(10, 19), link(11, 2), link(11, 7), link(11, 19), link(12, 6), link(12, 13), link(13, 0), link(13, 4), link(13, 6), link(13, 10), link(13, 12), link(13, 15), link(14, 10), link(15, 0), link(15, 3), link(15, 5), link(15, 6), link(15, 9), link(15, 13), link(15, 16), link(15, 18), link(16, 3), link(16, 15), link(17, 3), link(17, 19), link(18, 0), link(18, 1), link(18, 8), link(18, 15), link(18, 19), link(19, 10), link(19, 11), link(19, 17), link(19, 18), node(0), node(1), node(2), node(3), node(4), node(5), node(6), node(7), node(8), node(9), node(10), node(11), node(12), node(13), node(14), node(15), node(16), node(17), node(18), node(19), red(1), red(8), red(10), red(11), red(12), red(15), red(17) }
+SATISFIABLE
+TIME-STATS	java -jar Alpha-bundled.jar -n 1 -i test-src/3col-20-38.asp	0.82	0
