+++ benchmarks/stratified_evaluation/test-results/locstrat-200.clingo.out
@@ -1,0 +1,13 @@
+clingo version 5.2.2
+Reading from test-src/locstrat-200.asp
+Solving...
+Answer: 1
+p(0) p(1) p(2) p(3) p(4) p(5) p(6) p(7) p(8) p(9) p(10) p(11) p(12) p(13) p(14) p(15) p(16) p(17) p(18) p(19) p(20) p(21) p(22) p(23) p(24) p(25) p(26) p(27) p(28) p(29) p(30) p(31) p(32) p(33) p(34) p(35) p(36) p(37) p(38) p(39) p(40) p(41) p(42) p(43) p(44) p(45) p(46) p(47) p(48) p(49) p(50) p(51) p(52) p(53) p(54) p(55) p(56) p(57) p(58) p(59) p(60) p(61) p(62) p(63) p(64) p(65) p(66) p(67) p(68) p(69) p(70) p(71) p(72) p(73) p(74) p(75) p(76) p(77) p(78) p(79) p(80) p(81) p(82) p(83) p(84) p(85) p(86) p(87) p(88) p(89) p(90) p(91) p(92) p(93) p(94) p(95) p(96) p(97) p(98) p(99) p(100) p(101) p(102) p(103) p(104) p(105) p(106) p(107) p(108) p(109) p(110) p(111) p(112) p(113) p(114) p(115) p(116) p(117) p(118) p(119) p(120) p(121) p(122) p(123) p(124) p(125) p(126) p(127) p(128) p(129) p(130) p(131) p(132) p(133) p(134) p(135) p(136) p(137) p(138) p(139) p(140) p(141) p(142) p(143) p(144) p(145) p(146) p(147) p(148) p(149) p(150) p(151) p(152) p(153) p(154) p(155) p(156) p(157) p(158) p(159) p(160) p(161) p(162) p(163) p(164) p(165) p(166) p(167) p(168) p(169) p(170) p(171) p(172) p(173) p(174) p(175) p(176) p(177) p(178) p(179) p(180) p(181) p(182) p(183) p(184) p(185) p(186) p(187) p(188) p(189) p(190) p(191) p(192) p(193) p(194) p(195) p(196) p(197) p(198) p(199) a(0) a(1) a(2) a(3) a(4) a(5) a(6) a(7) a(8) a(9) a(10) a(11) a(12) a(13) a(14) a(15) a(16) a(17) a(18) a(19) a(20) a(21) a(22) a(23) a(24) a(25) a(26) a(27) a(28) a(29) a(30) a(31) a(32) a(33) a(34) a(35) a(36) a(37) a(38) a(39) a(40) a(41) a(42) a(43) a(44) a(45) a(46) a(47) a(48) a(49) a(50) a(51) a(52) a(53) a(54) a(55) a(56) a(57) a(58) a(59) a(60) a(61) a(62) a(63) a(64) a(65) a(66) a(67) a(68) a(69) a(70) a(71) a(72) a(73) a(74) a(75) a(76) a(77) a(78) a(79) a(80) a(81) a(82) a(83) a(84) a(85) a(86) a(87) a(88) a(89) a(90) a(91) a(92) a(93) a(94) a(95) a(96) a(97) a(98) a(99) a(100) a(101) a(102) a(103) a(104) a(105) a(106) a(107) a(108) a(109) a(110) a(111) a(112) a(113) a(114) a(115) a(116) a(117) a(118) a(119) a(120) a(121) a(122) a(123) a(124) a(125) a(126) a(127) a(128) a(129) a(130) a(131) a(132) a(133) a(134) a(135) a(136) a(137) a(138) a(139) a(140) a(141) a(142) a(143) a(144) a(145) a(146) a(147) a(148) a(149) a(150) a(151) a(152) a(153) a(154) a(155) a(156) a(157) a(158) a(159) a(160) a(161) a(162) a(163) a(164) a(165) a(166) a(167) a(168) a(169) a(170) a(171) a(172) a(173) a(174) a(175) a(176) a(177) a(178) a(179) a(180) a(181) a(182) a(183) a(184) a(185) a(186) a(187) a(188) a(189) a(190) a(191) a(192) a(193) a(194) a(195) a(196) a(197) a(198) a(199)
+SATISFIABLE
+
+Models       : 1+
+Calls        : 1
+Time         : 0.350s (Solving: 0.01s 1st Model: 0.01s Unsat: 0.00s)
+CPU Time     : 0.349s
+Command exited with non-zero status 10
+TIME-STATS	clingo -n 1 test-src/locstrat-200.asp	0.36	10
