+++ benchmarks/stratified_evaluation/test-results/locstrat-400.clingo.out
@@ -1,0 +1,13 @@
+clingo version 5.2.2
+Reading from test-src/locstrat-400.asp
+Solving...
+Answer: 1
+p(0) p(1) p(2) p(3) p(4) p(5) p(6) p(7) p(8) p(9) p(10) p(11) p(12) p(13) p(14) p(15) p(16) p(17) p(18) p(19) p(20) p(21) p(22) p(23) p(24) p(25) p(26) p(27) p(28) p(29) p(30) p(31) p(32) p(33) p(34) p(35) p(36) p(37) p(38) p(39) p(40) p(41) p(42) p(43) p(44) p(45) p(46) p(47) p(48) p(49) p(50) p(51) p(52) p(53) p(54) p(55) p(56) p(57) p(58) p(59) p(60) p(61) p(62) p(63) p(64) p(65) p(66) p(67) p(68) p(69) p(70) p(71) p(72) p(73) p(74) p(75) p(76) p(77) p(78) p(79) p(80) p(81) p(82) p(83) p(84) p(85) p(86) p(87) p(88) p(89) p(90) p(91) p(92) p(93) p(94) p(95) p(96) p(97) p(98) p(99) p(100) p(101) p(102) p(103) p(104) p(105) p(106) p(107) p(108) p(109) p(110) p(111) p(112) p(113) p(114) p(115) p(116) p(117) p(118) p(119) p(120) p(121) p(122) p(123) p(124) p(125) p(126) p(127) p(128) p(129) p(130) p(131) p(132) p(133) p(134) p(135) p(136) p(137) p(138) p(139) p(140) p(141) p(142) p(143) p(144) p(145) p(146) p(147) p(148) p(149) p(150) p(151) p(152) p(153) p(154) p(155) p(156) p(157) p(158) p(159) p(160) p(161) p(162) p(163) p(164) p(165) p(166) p(167) p(168) p(169) p(170) p(171) p(172) p(173) p(174) p(175) p(176) p(177) p(178) p(179) p(180) p(181) p(182) p(183) p(184) p(185) p(186) p(187) p(188) p(189) p(190) p(191) p(192) p(193) p(194) p(195) p(196) p(197) p(198) p(199) p(200) p(201) p(202) p(203) p(204) p(205) p(206) p(207) p(208) p(209) p(210) p(211) p(212) p(213) p(214) p(215) p(216) p(217) p(218) p(219) p(220) p(221) p(222) p(223) p(224) p(225) p(226) p(227) p(228) p(229) p(230) p(231) p(232) p(233) p(234) p(235) p(236) p(237) p(238) p(239) p(240) p(241) p(242) p(243) p(244) p(245) p(246) p(247) p(248) p(249) p(250) p(251) p(252) p(253) p(254) p(255) p(256) p(257) p(258) p(259) p(260) p(261) p(262) p(263) p(264) p(265) p(266) p(267) p(268) p(269) p(270) p(271) p(272) p(273) p(274) p(275) p(276) p(277) p(278) p(279) p(280) p(281) p(282) p(283) p(284) p(285) p(286) p(287) p(288) p(289) p(290) p(291) p(292) p(293) p(294) p(295) p(296) p(297) p(298) p(299) p(300) p(301) p(302) p(303) p(304) p(305) p(306) p(307) p(308) p(309) p(310) p(311) p(312) p(313) p(314) p(315) p(316) p(317) p(318) p(319) p(320) p(321) p(322) p(323) p(324) p(325) p(326) p(327) p(328) p(329) p(330) p(331) p(332) p(333) p(334) p(335) p(336) p(337) p(338) p(339) p(340) p(341) p(342) p(343) p(344) p(345) p(346) p(347) p(348) p(349) p(350) p(351) p(352) p(353) p(354) p(355) p(356) p(357) p(358) p(359) p(360) p(361) p(362) p(363) p(364) p(365) p(366) p(367) p(368) p(369) p(370) p(371) p(372) p(373) p(374) p(375) p(376) p(377) p(378) p(379) p(380) p(381) p(382) p(383) p(384) p(385) p(386) p(387) p(388) p(389) p(390) p(391) p(392) p(393) p(394) p(395) p(396) p(397) p(398) p(399) a(0) a(1) a(2) a(3) a(4) a(5) a(6) a(7) a(8) a(9) a(10) a(11) a(12) a(13) a(14) a(15) a(16) a(17) a(18) a(19) a(20) a(21) a(22) a(23) a(24) a(25) a(26) a(27) a(28) a(29) a(30) a(31) a(32) a(33) a(34) a(35) a(36) a(37) a(38) a(39) a(40) a(41) a(42) a(43) a(44) a(45) a(46) a(47) a(48) a(49) a(50) a(51) a(52) a(53) a(54) a(55) a(56) a(57) a(58) a(59) a(60) a(61) a(62) a(63) a(64) a(65) a(66) a(67) a(68) a(69) a(70) a(71) a(72) a(73) a(74) a(75) a(76) a(77) a(78) a(79) a(80) a(81) a(82) a(83) a(84) a(85) a(86) a(87) a(88) a(89) a(90) a(91) a(92) a(93) a(94) a(95) a(96) a(97) a(98) a(99) a(100) a(101) a(102) a(103) a(104) a(105) a(106) a(107) a(108) a(109) a(110) a(111) a(112) a(113) a(114) a(115) a(116) a(117) a(118) a(119) a(120) a(121) a(122) a(123) a(124) a(125) a(126) a(127) a(128) a(129) a(130) a(131) a(132) a(133) a(134) a(135) a(136) a(137) a(138) a(139) a(140) a(141) a(142) a(143) a(144) a(145) a(146) a(147) a(148) a(149) a(150) a(151) a(152) a(153) a(154) a(155) a(156) a(157) a(158) a(159) a(160) a(161) a(162) a(163) a(164) a(165) a(166) a(167) a(168) a(169) a(170) a(171) a(172) a(173) a(174) a(175) a(176) a(177) a(178) a(179) a(180) a(181) a(182) a(183) a(184) a(185) a(186) a(187) a(188) a(189) a(190) a(191) a(192) a(193) a(194) a(195) a(196) a(197) a(198) a(199) a(200) a(201) a(202) a(203) a(204) a(205) a(206) a(207) a(208) a(209) a(210) a(211) a(212) a(213) a(214) a(215) a(216) a(217) a(218) a(219) a(220) a(221) a(222) a(223) a(224) a(225) a(226) a(227) a(228) a(229) a(230) a(231) a(232) a(233) a(234) a(235) a(236) a(237) a(238) a(239) a(240) a(241) a(242) a(243) a(244) a(245) a(246) a(247) a(248) a(249) a(250) a(251) a(252) a(253) a(254) a(255) a(256) a(257) a(258) a(259) a(260) a(261) a(262) a(263) a(264) a(265) a(266) a(267) a(268) a(269) a(270) a(271) a(272) a(273) a(274) a(275) a(276) a(277) a(278) a(279) a(280) a(281) a(282) a(283) a(284) a(285) a(286) a(287) a(288) a(289) a(290) a(291) a(292) a(293) a(294) a(295) a(296) a(297) a(298) a(299) a(300) a(301) a(302) a(303) a(304) a(305) a(306) a(307) a(308) a(309) a(310) a(311) a(312) a(313) a(314) a(315) a(316) a(317) a(318) a(319) a(320) a(321) a(322) a(323) a(324) a(325) a(326) a(327) a(328) a(329) a(330) a(331) a(332) a(333) a(334) a(335) a(336) a(337) a(338) a(339) a(340) a(341) a(342) a(343) a(344) a(345) a(346) a(347) a(348) a(349) a(350) a(351) a(352) a(353) a(354) a(355) a(356) a(357) a(358) a(359) a(360) a(361) a(362) a(363) a(364) a(365) a(366) a(367) a(368) a(369) a(370) a(371) a(372) a(373) a(374) a(375) a(376) a(377) a(378) a(379) a(380) a(381) a(382) a(383) a(384) a(385) a(386) a(387) a(388) a(389) a(390) a(391) a(392) a(393) a(394) a(395) a(396) a(397) a(398) a(399)
+SATISFIABLE
+
+Models       : 1+
+Calls        : 1
+Time         : 1.617s (Solving: 0.03s 1st Model: 0.03s Unsat: 0.00s)
+CPU Time     : 1.603s
+Command exited with non-zero status 10
+TIME-STATS	clingo -n 1 test-src/locstrat-400.asp	1.94	10
