+++ benchmarks/stratified_evaluation/test-results/reach-4.alpha_nse.as
@@ -1,0 +1,3 @@
+	at at.ac.tuwien.kr.alpha.Main.main(Main.java:124)
+Command exited with non-zero status 1
+TIME-STATS	java -jar Alpha-bundled.jar -nse -n 1 -i test-src/reach-4.asp	91.13	1
