+++ src/main/java/at/ac/tuwien/kr/alpha/Alpha.java
@@ -58,8 +58,8 @@ import at.ac.tuwien.kr.alpha.config.SystemConfig;
 import at.ac.tuwien.kr.alpha.grounder.Grounder;
 import at.ac.tuwien.kr.alpha.grounder.GrounderFactory;
 import at.ac.tuwien.kr.alpha.grounder.parser.ProgramParser;
+import at.ac.tuwien.kr.alpha.grounder.transformation.eval.PartialEvaluation;
 import at.ac.tuwien.kr.alpha.grounder.transformation.impl.NormalizeProgramTransformation;
-import at.ac.tuwien.kr.alpha.grounder.transformation.impl.PartialEvaluation;
 import at.ac.tuwien.kr.alpha.solver.Solver;
 import at.ac.tuwien.kr.alpha.solver.SolverFactory;
 import at.ac.tuwien.kr.alpha.solver.heuristics.HeuristicsConfiguration;
